[Agent] Extract token override mixin
