+++ tests/common/containerTestBed.js
@@ -4,19 +4,16 @@
 
 import { jest } from '@jest/globals';
 import BaseTestBed from './baseTestBed.js';
+import { TokenOverrideMixin } from './tokenOverrideMixin.js';
 
 /**
  * @description Base class that manages a DI container and allows overriding
  * token resolution during tests.
  * @class
  */
-export class ContainerTestBed extends BaseTestBed {
+export class ContainerTestBed extends TokenOverrideMixin(BaseTestBed) {
   /** @type {{ resolve: jest.Mock }} */
   container;
-  /** @type {Map<any, any>} */
-  #tokenOverrides = new Map();
-  /** @type {Function} */
-  #originalResolve;
 
   /**
    * @description Constructs the test bed with the provided container.
@@ -26,49 +23,7 @@ export class ContainerTestBed extends BaseTestBed {
   constructor(container, mocks = {}) {
     super(mocks);
     this.container = container;
-    this.#originalResolve =
-      this.container.resolve.getMockImplementation?.() ??
-      this.container.resolve;
-  }
-
-  /**
-   * Temporarily overrides container token resolution.
-   *
-   * @param {any} token - Token to override.
-   * @param {any | (() => any)} value - Replacement value or function.
-   * @returns {void}
-   */
-  withTokenOverride(token, value) {
-    this.#tokenOverrides.set(token, value);
-    this.container.resolve.mockImplementation((tok) => {
-      if (this.#tokenOverrides.has(tok)) {
-        const override = this.#tokenOverrides.get(tok);
-        return typeof override === 'function' ? override() : override;
-      }
-      return this.#originalResolve(tok);
-    });
-  }
-
-  /**
-   * Restores the container state and clears overrides.
-   *
-   * @returns {Promise<void>} Promise resolving when cleanup is complete.
-   */
-  async cleanup() {
-    await super.cleanup();
-  }
-
-  /**
-   * Restores the container state and clears overrides after base cleanup.
-   *
-   * @protected
-   * @returns {Promise<void>} Promise resolving when container cleanup is complete.
-   */
-  async _afterCleanup() {
-    this.container.resolve.mockImplementation(this.#originalResolve);
-    this.#tokenOverrides.clear();
-    await super._afterCleanup();
+    this._initTokenOverrides(container);
   }
 }
-
 export default ContainerTestBed;
